vue_front_end_code/login-widget/requirements.txt
streamlit
GitPython
llama-cpp-python
pytest
ollama
